moving AuthContext.tsx
