+++ routecarboniq/app/frontend/dashboard/page.tsx
@@ -2,7 +2,7 @@
 
 import { useEffect } from "react";
 import { useRouter } from "next/navigation";
-import { useAuth } from "../src/context/AuthContext";
+import { useAuth } from "../context/AuthContext";
 
 export default function DashboardPage() {
   const { user, loading } = useAuth();
