+++ routecarboniq/app/frontend/stm/page.tsx
@@ -2,7 +2,7 @@
 
 import {useEffect, useState} from "react";
 import { useRouter } from "next/navigation";
-import { useAuth } from "../src/context/AuthContext";
+import { useAuth } from "../context/AuthContext";
 import { Leaf} from "lucide-react";
 import Form from 'next/form'
